requirements.txt
Django==4.2.3
django-admin-autocomplete-filter==0.7.1
django-admin-rangefilter==0.10.0
django-import-export==3.2.0
